added authorization headers
