+++ client/src/components/Login/Login.scss
@@ -23,6 +23,7 @@
     .password-input {
       color: white;
       width: 250px;
+      text-align: start;
       input {
         color: white;
       }
