+++ client/src/components/Signout/Signout.tsx
@@ -1,6 +1,7 @@
 import React, { Dispatch, SetStateAction, useEffect } from "react";
 import { useHistory } from "react-router";
 import { Ctx } from "../../App";
+import { setAccessToken } from "../../util/Token";
 
 interface Props {
   setToken: Dispatch<SetStateAction<string>>;
@@ -12,7 +13,7 @@ export const Signout: React.FC<Props> = props => {
 
   useEffect(() => {
     ctx.user = undefined;
-    ctx.setAccessToken("");
+    setAccessToken("");
     props.setToken("");
     history.push("/");
   }, []);
