+++ client/src/util/Token.ts
@@ -1,0 +1,7 @@
+export const getAccessToken: () => string = () => {
+  return sessionStorage.getItem("accessToken") ?? "";
+};
+
+export const setAccessToken: (token: string) => void = (token: string) => {
+  sessionStorage.setItem("accessToken", token);
+};
