+++ server/src/routes/password.ts
@@ -8,7 +8,7 @@ export default function getRoutes(server: Hapi.Server): void {
     method: "PUT",
     path: "/passwordreset/{userId}",
     options: {
-      auth: false,
+      auth: "jwt",
       validate: {
         params: {
           userId: Joi.string().required()
