+++ server/src/routes/users.ts
@@ -8,8 +8,7 @@ export default function getRoutes(server: Hapi.Server): void {
     method: "GET",
     path: "/user/{userId}",
     options: {
-      auth: false,
-      // auth: "jwt",
+      auth: "jwt",
       validate: {
         params: {
           userId: Joi.string().required()
@@ -24,8 +23,7 @@ export default function getRoutes(server: Hapi.Server): void {
     method: "GET",
     path: "/user",
     options: {
-      auth: false
-      // auth: "jwt"
+      auth: "jwt"
     },
     handler: UserController.list
   });
@@ -34,8 +32,7 @@ export default function getRoutes(server: Hapi.Server): void {
     method: "DELETE",
     path: "/user/{userId}",
     options: {
-      auth: false,
-      // auth: "jwt",
+      auth: "jwt",
       validate: {
         params: {
           userId: Joi.string().required()
@@ -50,8 +47,7 @@ export default function getRoutes(server: Hapi.Server): void {
     method: "PUT",
     path: "/user/{userId}",
     options: {
-      auth: false,
-      // auth: "jwt",
+      auth: "jwt",
       validate: {
         params: {
           userId: Joi.string().required()
